Improved testing for batch metrics.
Refactored the test to include assertion metrics for counter batches
in addition to what was previously tested in logged and unlogged batches.
Modified tests to assert random ranges of batches, partitions and statements,
printing the seed for the Random on failure so that the error state could be
recreated.

Patch by Stephen Mallette; Reviewed by Jordan West and Dinesh Joshi for CASSANDRA-15718